Create initial .travis.yml for macOS/Linux
This uses cibuildwheel to generate binary wheels for macOS and Linux.
cibuildwheel can also work for Windows, but I'll leave this for now as I have
not worked with appveyor.

Installation is based on the "Ubuntu Install Experience" wiki page. Dynamic
linkage seems to be ignored; mupdf appears to package everything it uses,
except for <zlib.h>, so setup.py is adjusted to fit the needs of this deploy-
ment rather than private use. It turns out, happily, that it is actually
unnecessary to modify any mupdf sources as those instructions recommended.

i686 is disabled because it is now only 1% of the Linux user base, and it did
not work. Python 3.3 is disabled because while supported by cibuildwheel, it
is past end of life.

https://github.com/joerick/cibuildwheel