fitz/utils.py: fix #2051, added `dpi` arg in get_page_pixmap() and require keyword-only args.
In `get_pixmap()`, use Python's special `*` arg to force args to be
keyword-only instead of enforcing this in the code.

In `get_page_pixmap()`, also use Python's special `*` arg to force args
to be keyword-only. This makes the fn conform to the documentation in
docs/document.rst, but could break code that used to use non-keyword args.