Fixing #2300
Page method `bound()` wraps MuPDF function fz_bound_page, but did not catch possible exceptions raised by this function. In these (rare) cases the Python interpreter will end. We now wrap the function with try / catch but do not raise an exception - instead the MuPDF error message is displayed. For non-PDF documents the infinite rectangle is returned, for PDFs the page rectangle is instead derived from the CropBox as fallback.