Adding PDF Names resolution
Add Document method "resolve_names()".
It walks through the "/Dests" and "/Names/Dests" dictionaries in a PDF's catalog and returns a Python dictionary with the name as key and value dictionary containing the target page and location.
Note that this method is only implemented in "rebased" and will not work in classic architecture.