pipcl.py: allow providing python-config externally
When cross-compiling (e.g. using Buildroot), the python-config
executable that resides next to the host python executable provides
incorrect includes (the ones for the host).

Since the correct path to python-config cannot be guessed, add an
additional environment variable to allow setting the path to the
correct python-config executable externally.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>