Fixes
Address bug 3207:
Multiple sub-paths in a path were not correctly handled when "closepath" commands were present.
When encountering "closepath", this fix generates a line that connects the current point with the first point of the current sub-path - however only if these two points are different.
The "firstpoint" value is being set exclusively by the "moveto" command.