Detect dotted gridlines for tables
Addresses #3539

We previously did not detect dotted lines used as table gridlines.

If one of width / height is LE edge_min_length and LT the other dimension, the rectangle is treated as a vertical / horizontal line.
We incorrectly used the dimension-specific snap values for this.

We now no longer ignore rectangles if both, width and height are smaller then edge_min_length, but leave this to the snapping and joining algorithms further down the road.